Player Movement And Debug System
I finished the debug system and implemented a rigidbody player movement script. This allows me to move in all dimensions along with crouching and sprinting in a smooth and effective manner.

For the debug system I made it so that it can be opened using keybinds whilst in game, and also added all of the player movement information to it