Modelling Progress
Begun modelling for the initial cutscene, made the terrain, materials and lighting setup. Made a path for the car to follow and animated that, using metaballs for smoke and particles for tire tracks in the sand.

Next I did several test renders fixing various issues and working out what quality is needed.

Now im working on the bean industries building, I have built the majority of the above ground section and am now working on the main structure of the building. Once everything is modelled I will begin the bean animations and plan the camera angles