Made weapons work realistically
Weapons now shoot realistically and accurately, zeroing the gun on the players line of sight if they can see what theyre shooting

raycast from player center of view

if raycast hits object, get hit location
use drawline from the gun fire point to see if it can reach that location unobstructed

if it can, fire from the players viewpoint.

This means that the players aim will always be true unless the weapon is obstructed, if it is obstructed the gun will just shoot whatever obstacle is in the way.