Target changes
After discovering targets would not swing in builds i began researching why.

In conclusion im still unsure however i narrowed it down to my method of moving them being to adjust the anchor point. Though this shouldve worked it for some reason doesent in builds which im assuming is a lack of documentation on unitys part. After further digging and experimenting I created a system using 3 hinge joints instead of 2, these are moved by moving the top one with rigidbody move.

This solves literally all of the problems and is amazing